+++ cnext_server/server/python/dataframe_manager/dataframe_manager.py
@@ -13,6 +13,7 @@ from libs.message_handler import BaseMessageHandler
 from libs.message import ContentType, DFManagerCommand, SubContentType
 from libs.json_serializable import ipython_internal_output, JsonSerializable
 from cnextlib.mime_types import CnextMimeType
+from .dataframe import DataFrame
 
 from libs import logs
 from libs.message import Message, WebappEndpoint
@@ -153,11 +154,12 @@ class MessageHandler(BaseMessageHandler):
         return tableData
 
     @ipython_internal_output
-    def _ipython_get_table_data(self, df_id, code):
+    def _ipython_get_table_data(self, df_id, df_type, filter, from_index, to_index):
         output = None
         self.user_space.execute(
             "print('Get table data for dataframe %s...')" % df_id, ExecutionMode.EVAL)
-        result = self.user_space.execute(code, ExecutionMode.EVAL)
+        dataframe = DataFrame(self.user_space, df_id, df_type)
+        result = dataframe.get_table_data(filter, from_index, to_index)
         # print("get table data %s" % result)
         # log.info("get table data %s" % result)
         if result is not None:
@@ -166,39 +168,18 @@ class MessageHandler(BaseMessageHandler):
         return output
 
     @ipython_internal_output
-    def _ipython_get_metadata(self, df_id):
-        self.user_space.execute("print('Get metadata for dataframe %s...')" % df_id, ExecutionMode.EVAL)
-        shape = self.user_space.execute("%s.shape" % df_id, ExecutionMode.EVAL)
-        df_type = self.user_space.execute(
-            "%s.__module__ + '.' + %s.__class__.__name__" % (df_id, df_id), ExecutionMode.EVAL)
-        dtypes = self.user_space.execute(
-            "%s.dtypes" % df_id, ExecutionMode.EVAL)
-        countna = self.user_space.execute(
-            "%s.isna().sum()" % df_id, ExecutionMode.EVAL)
-        describe = self.user_space.execute(
-            "%s.describe(include='all')" % df_id, ExecutionMode.EVAL)
-        describe = self._convert_to_str_if_not_jsonable(describe)
-
-        columns = {}
-        MAX_UNIQUE_LENGTH = 1000
-        for col_name, ctype in dtypes.items():
-            if re.search(r'datetime', ctype.name):
-                # the unique value of datetime is usually the same as the length so it is meaningless
-                # also we have to convert it to string before sending back. So just don't do it now.
-                unique = []
-            else:
-                unique = self.user_space.execute(
-                    "_pd.Series(%s['%s'].unique()).tolist()" % (df_id, col_name), ExecutionMode.EVAL)
-
-            # only send unique list that has length smaller than MAX_UNIQUE_LENGTH
-            if len(unique) > MAX_UNIQUE_LENGTH:
-                unique = []
-
-            columns[col_name] = {'name': col_name,
-                                 'type': str(ctype.name), 'unique': unique,
-                                 'countna': countna[col_name].item(), 'describe': describe[col_name].to_dict()}
+    def _ipython_get_metadata(self, df_id, df_type):
+        self.user_space.execute(
+            "print('Get metadata for dataframe %s...')" % df_id, ExecutionMode.EVAL)
+        dataframe = DataFrame(self.user_space, df_id, df_type)
+        shape, dtypes, countna, describe, nuniques = dataframe.get_metadata()
+        uniques = dataframe.uniques(df_id, dtypes, nuniques)
+        columns = dataframe.get_column_summary(
+            dtypes, countna, describe, uniques)
         output = {'df_id': df_id, 'type': str(df_type),
                   'shape': shape, 'columns': columns, 'timestamp': time.time()}
+        self.user_space.execute(
+            "print('Done!')", ExecutionMode.EVAL)
         return output
 
     # this function is run inside ipython but we don't have to wrap it with ipython_internal_output
@@ -311,17 +292,24 @@ class MessageHandler(BaseMessageHandler):
         ## this step is important to make sure the query work properly in the backend #
         if (isinstance(message.content, str)):
             message.content = message.content.replace("'", '"')
-
+        for key in message.metadata:
+            if (isinstance(message.metadata[key], str)):
+                message.metadata[key] = message.metadata[key].replace("'", '"')
+                
         try:
             if self.user_space.is_alive():
                 if message.command_name == DFManagerCommand.get_table_data:
                     # TODO: turn _df_manager to variable
-                    self.user_space.execute("{}._ipython_get_table_data('{}', '{}')".format(
-                        IPythonInteral.DF_MANAGER.value, message.metadata['df_id'], message.content), ExecutionMode.EVAL, self.message_handler_callback, message)
+                    self.user_space.execute("{}._ipython_get_table_data('{}', '{}', '{}', '{}', '{}')".format(
+                        IPythonInteral.DF_MANAGER.value, message.metadata['df_id'], message.metadata['df_type'], 
+                        message.metadata['filter'] if message.metadata['filter'] is not None else "", 
+                        message.metadata['from_index'], message.metadata['to_index']), 
+                        ExecutionMode.EVAL, self.message_handler_callback, message)
 
                 elif message.command_name == DFManagerCommand.get_df_metadata:
-                    self.user_space.execute("{}._ipython_get_metadata('{}')".format(
-                        IPythonInteral.DF_MANAGER.value, message.metadata['df_id']), ExecutionMode.EVAL, self.message_handler_callback, message)
+                    self.user_space.execute("{}._ipython_get_metadata('{}', '{}')".format(
+                        IPythonInteral.DF_MANAGER.value, message.metadata['df_id'], message.metadata['df_type']),
+                         ExecutionMode.EVAL, self.message_handler_callback, message)
 
                 elif message.command_name == DFManagerCommand.get_registered_udfs:
                     self.user_space.execute("{}._ipython_get_registered_udfs()".format(